Move Input and Output inside ReactiveBlock, since they are created for and by the block
